refactor: Remove cart reducers
