+++ projects/core/src/cart/cart.module.ts
@@ -10,7 +10,6 @@ import {
 import { MultiCartService } from './facade/multi-cart.service';
 import { CartPageMetaResolver } from './services/cart-page-meta.resolver';
 import { MultiCartStatePersistenceService } from './services/multi-cart-state-persistence.service';
-import { CartStoreModule } from './store/cart-store.module';
 import { MultiCartStoreModule } from './store/multi-cart-store.module';
 
 export function cartStatePersistenceFactory(
@@ -25,7 +24,7 @@ export function cartStatePersistenceFactory(
 }
 
 @NgModule({
-  imports: [CartStoreModule, MultiCartStoreModule],
+  imports: [MultiCartStoreModule],
 })
 export class CartModule {
   static forRoot(): ModuleWithProviders<CartModule> {
